Create a nested 'kgames' menu for the desktop
This places all of the kgames applications inside a 'KGames' menu
which lives in the main Games menu.

This adds a configuration option as to whether to install the .menu
file in the current user's home directory -- there isn't a location in
/usr/local which is defined for this file, only /etc or $HOME. This is
enabled by default so that a 'ninja install' works and doesn't require
admin privs. Switch back to the /etc location by doing

	$ meson -Duser-menu=false

Signed-off-by: Keith Packard <keithp@keithp.com>