kcribbage: Don't run past the end of the deckCards array in updateCards
updateCards wants to remove unused cards from the GUI, so it walks
over the rest of the cards array checking for cards in use. For
deckCards, the array is only 1 long, but the length used was NUM_CARDS
(10), meaning that this function was walking 9 cards past the end of
the array and potentially smashing bits of the heap.

Signed-off-by: Keith Packard <keithp@keithp.com>